backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database (MongoDB)
pymongo==4.4.1

# Geospatial
shapely==2.0.6

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Geospatial data processing
rasterio==1.4.2
torchgeo==0.6.0
torch==2.5.1
torchvision==0.20.1

# Climate/weather data
xarray==2024.10.0
netCDF4==1.7.2
h5netcdf==1.3.0

# HTTP client
httpx==0.27.2

# Data processing
pandas==2.2.3
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12

# Additional optional dependencies requested by user
# HTTP client
requests==2.31.0
# Google API / Drive support
google-api-python-client==2.95.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
# OAuth helper for Google APIs (provides `google_auth_oauthlib`)
google-auth-oauthlib==1.0.0
# For Windows support of curses-based libraries (only on Windows)
# Use a compatible windows-curses release (avoid strict pin to 2.3.0)
windows-curses>=2.3.1,<2.5.0
